(refactor) Move typescript code to src.
