+++ bower.json
@@ -18,6 +18,9 @@
     "app-layout": "PolymerElements/app-layout#^2.0.0",
     "font-roboto": "PolymerElements/font-roboto-local#^1.0.0",
     "iron-component-page": "PolymerElements/iron-component-page#^3.0.0",
+    "iron-collapse": "PolymerElements/iron-collapse#^2.0.0",
+    "iron-list": "PolymerElements/iron-list#^2.0.0",
+    "paper-item": "PolymerElements/paper-item#^2.0.0",
     "paper-progress": "PolymerElements/paper-progress#^2.0.0",
     "paper-spinner": "PolymerElements/paper-spinner#^2.0.0",
     "web-component-tester": "Polymer/web-component-tester#6.5.0",
