+++ oo-worker.html
@@ -39,8 +39,8 @@
           super.ready()
           console.log(oo)
 
-          const worker = new oo.Worker()
-          console.log(worker)
+          this._worker = new oo.Worker()
+          console.log(this._worker)
         }
       }
