+++ package.json
@@ -33,10 +33,11 @@
   "scripts": {
     "browser": "./node_modules/.bin/npm-run-all --parallel serve browser:after",
     "browser:after": "./node_modules/.bin/npm-run-all --serial util:delay serve:watch",
-    "build": "./node_modules/.bin/polymer build",
+    "build": "./node_modules/.bin/npm-run-all --serial build:node build:polymer",
     "build:docs": "./node_modules/.bin/polymer analyze > analysis.json",
-    "bundle": "./node_modules/.bin/browserify oo-worker.ts --plugin [ tsify --module=commonjs ] --standalone oo --outfile oo-worker.dist.js",
-    "bundle:watch": "./node_modules/.bin/watchify oo-worker.ts --plugin [ tsify --module=commonjs ] --standalone oo --outfile oo-worker.dist.js",
+    "build:node": "./node_modules/.bin/browserify src/oo-worker.ts --plugin [ tsify --module=commonjs ] --standalone oo --outfile oo-worker.dist.js",
+    "build:node:watch": "./node_modules/.bin/watchify src/oo-worker.ts --plugin [ tsify --module=commonjs ] --standalone oo --outfile oo-worker.dist.js",
+    "build:polymer": "./node_modules/.bin/polymer build",
     "lint": "npm run lint:polymer",
     "lint:js": "./node_modules/.bin/eslint . --ext js,html --ignore-path .gitignore",
     "lint:polymer": "./node_modules/.bin/polymer lint",
