+++ src/tinyqueue.d.ts
